adding new dataloader with basic data augmentation capabilities
the new dataloader does not load BEVs but generates them using the
results from AANET + ALVAROMASK (DualBisnet) and the Q-Matrix obtained
with StereoRectify. Currently the Q-Matrix is hard coded, would be nice
to have this matrix evaluated once in the code for both the days.

Data augmentation is done by
 1) adding random noise on the ROTATION
 2) adding random noise on the TRANSLATION
 see the documentation inside the code to understand how the image is
 transformed; values for noise generation are currently hard coded.

Warning: the Q-Matrix is the same for both 2011_09_30 and 2011_10_03.